app/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.30.0
gradio>=3.50.0
Pillow>=8.3.1
requests>=2.31.0 